requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.3
django-channels==0.7.0
django-cors-headers==3.11.0
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2022.1
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.0
zope.interface==5.4.0